requirements.txt
fastapi
uvicorn[standard]
python-multipart

pdfplumber
pdf2image
pytesseract
Pillow

httpx
pdfminer.six==20250324
